renamed tuple_formatter module
