+++ src/rt_core_v2/rttuple_formatter.py
